Rename scripts and update references
